+++ Docs/API/html/search/functions_0.js
@@ -3,5 +3,5 @@ var searchData=
   ['callfunction',['CallFunction',['../class_jinx_1_1_i_script.html#af4df3369ab82192a9a00c7eae47d7e15',1,'Jinx::IScript']]],
   ['compile',['Compile',['../class_jinx_1_1_i_runtime.html#a85a3a6697eea6c30b8d02b4ca1dc2131',1,'Jinx::IRuntime']]],
   ['createruntime',['CreateRuntime',['../_jinx_8h.html#a5016b82d1aaae052ed18a98aef3af7fa',1,'Jinx']]],
-  ['createscript',['CreateScript',['../class_jinx_1_1_i_runtime.html#a82be192865ed4378221c93abc01eb3b9',1,'Jinx::IRuntime::CreateScript(BufferPtr bytecode, std::any userContext=nullptr)=0'],['../class_jinx_1_1_i_runtime.html#aa701adaee9d308dc3662d5dbc8869a0a',1,'Jinx::IRuntime::CreateScript(const char *scriptText, std::any userContext=nullptr, String name=String(), std::initializer_list&lt; String &gt; libraries={})=0']]]
+  ['createscript',['CreateScript',['../class_jinx_1_1_i_runtime.html#a239db2d9140dca0ec8ae96d3b178e871',1,'Jinx::IRuntime::CreateScript(BufferPtr bytecode, Any userContext=nullptr)=0'],['../class_jinx_1_1_i_runtime.html#a6c79daf99b262418ff7e074696f9c76b',1,'Jinx::IRuntime::CreateScript(const char *scriptText, Any userContext=nullptr, String name=String(), std::initializer_list&lt; String &gt; libraries={})=0']]]
 ];