+++ Include/Jinx.hpp
@@ -86,6 +86,14 @@ by James Boer, and distributed under the MIT license.
 #endif
 #endif
 
+/*
+On macOS, use of std::any is restricted to applications targeting versions 10.14 and up (Mohave) due to
+limitations in std::any_cast.  As such, Jinx provides optional void * aliases in place of std::any in
+case a project wishes to target macOS clients earlier than 10.14.
+*/
+
+#define JINX_USE_ANY
+
 #include <memory>
 #include <functional>
 #include <vector>
@@ -94,7 +102,9 @@ by James Boer, and distributed under the MIT license.
 #include <cstddef>
 #include <limits>
 #include <cstring>
+#ifdef JINX_USE_ANY
 #include <any>
+#endif
 
 #ifdef JINX_WINDOWS
 #pragma warning(pop)
@@ -699,6 +709,16 @@ namespace Jinx
 	using RuntimeID = uint64_t;
 	const RuntimeID InvalidID = 0;
 
+#ifdef JINX_USE_ANY
+	using Any = std::any;
+#define JinxAny std::any
+#define JinxAnyCast std::any_cast
+#else
+	using Any = void *;
+#define JinxAny Jinx::Any
+#define JinxAnyCast reinterpret_cast
+#endif
+
 	enum class Visibility
 	{
 		Public,
@@ -852,7 +872,7 @@ namespace Jinx
 		\return void pointer optionally passed at script creation.  This is intended to be
 		used by native library functions to retrieve user-specific data or objects.
 		*/
-		virtual std::any GetUserContext() const = 0;
+		virtual Any GetUserContext() const = 0;
 
 		/// Return the library to which this script belongs
 		/**
@@ -929,7 +949,7 @@ namespace Jinx
 		\return A ScriptPtr ready for execution.
 		\sa Compile(), IScript
 		*/
-		virtual ScriptPtr CreateScript(BufferPtr bytecode, std::any userContext = nullptr) = 0;
+		virtual ScriptPtr CreateScript(BufferPtr bytecode, Any userContext = nullptr) = 0;
 
 		/// Compile and create script from text
 		/**
@@ -941,7 +961,7 @@ namespace Jinx
 		\param libraries A list of libraries to import by default.
 		\return A ScriptPtr containing compiled bytecode on success or a nullptr on failure.
 		*/
-		virtual ScriptPtr CreateScript(const char * scriptText, std::any userContext = nullptr, String name = String(), std::initializer_list<String> libraries = {}) = 0;
+		virtual ScriptPtr CreateScript(const char * scriptText, Any userContext = nullptr, String name = String(), std::initializer_list<String> libraries = {}) = 0;
 
 		/// Compile, create, and execute a script
 		/**
@@ -954,7 +974,7 @@ namespace Jinx
 		\param libraries A list of libraries to import by default.
 		\return A ScriptPtr containing compiled bytecode on success or a nullptr on failure.
 		*/
-		virtual ScriptPtr ExecuteScript(const char * scriptText, std::any userContext = nullptr, String name = String(), std::initializer_list<String> libraries = {}) = 0;
+		virtual ScriptPtr ExecuteScript(const char * scriptText, Any userContext = nullptr, String name = String(), std::initializer_list<String> libraries = {}) = 0;
 
 		/// Retrieve library by name or create empty library if not found
 		/**
@@ -2473,7 +2493,7 @@ namespace Jinx::Impl
 	class Script : public IScript, public std::enable_shared_from_this<Script>
 	{
 	public:
-		Script(RuntimeIPtr runtime, BufferPtr bytecode, std::any userContext);
+		Script(RuntimeIPtr runtime, BufferPtr bytecode, Any userContext);
 		virtual ~Script();
 
 		bool RegisterFunction(LibraryPtr library, Visibility visibility, const String & name, FunctionCallback function) override;
@@ -2487,7 +2507,7 @@ namespace Jinx::Impl
 		void SetVariable(const String & name, const Variant & value) override;
 
 		const String & GetName() const override { return m_name; }
-		std::any GetUserContext() const override { return m_userContext; }
+		Any GetUserContext() const override { return m_userContext; }
 		LibraryPtr GetLibrary() const override { return m_library; }
 
 		std::vector<String, Allocator<String>> GetCallStack() const;
@@ -2559,7 +2579,7 @@ namespace Jinx::Impl
 		FunctionMap m_functionMap;
 
 		// User context pointer
-		std::any m_userContext;
+		Any m_userContext;
 
 		// Initial position of bytecode for this script
 		size_t m_bytecodeStart;
@@ -2610,9 +2630,9 @@ namespace Jinx::Impl
 
 		// IRuntime interface
 		BufferPtr Compile(const char * scriptText, String name, std::initializer_list<String> libraries) override;
-		ScriptPtr CreateScript(BufferPtr bytecode, std::any userContext) override;
-		ScriptPtr CreateScript(const char * scriptText, std::any userContext, String name, std::initializer_list<String> libraries) override;
-		ScriptPtr ExecuteScript(const char * scriptText, std::any userContext, String name, std::initializer_list<String> libraries) override;
+		ScriptPtr CreateScript(BufferPtr bytecode, Any userContext) override;
+		ScriptPtr CreateScript(const char * scriptText, Any userContext, String name, std::initializer_list<String> libraries) override;
+		ScriptPtr ExecuteScript(const char * scriptText, Any userContext, String name, std::initializer_list<String> libraries) override;
 		LibraryPtr GetLibrary(const String & name) override;
 		PerformanceStats GetScriptPerformanceStats(bool resetStats = true) override;
 		BufferPtr StripDebugInfo(BufferPtr bytecode) const override;
@@ -8170,12 +8190,12 @@ namespace Jinx::Impl
 		return Compile(scriptBuffer, name, libraries);
 	}
 
-	inline ScriptPtr Runtime::CreateScript(BufferPtr bytecode, std::any userContext)
+	inline ScriptPtr Runtime::CreateScript(BufferPtr bytecode, Any userContext)
 	{
 		return std::allocate_shared<Script>(Allocator<Script>(), shared_from_this(), std::static_pointer_cast<Buffer>(bytecode), userContext);
 	}
 
-	inline ScriptPtr Runtime::CreateScript(const char * scriptText, std::any userContext, String name, std::initializer_list<String> libraries)
+	inline ScriptPtr Runtime::CreateScript(const char * scriptText, Any userContext, String name, std::initializer_list<String> libraries)
 	{
 		// Compile script text to bytecode
 		auto bytecode = Compile(scriptText, name, libraries);
@@ -8186,7 +8206,7 @@ namespace Jinx::Impl
 		return CreateScript(bytecode, userContext);
 	}
 
-	inline ScriptPtr Runtime::ExecuteScript(const char * scriptcode, std::any userContext, String name, std::initializer_list<String> libraries)
+	inline ScriptPtr Runtime::ExecuteScript(const char * scriptcode, Any userContext, String name, std::initializer_list<String> libraries)
 	{
 		// Compile the text to bytecode
 		auto bytecode = Compile(scriptcode, name, libraries);
@@ -8565,7 +8585,7 @@ Copyright (c) 2016 James Boer
 namespace Jinx::Impl
 {
 
-	inline Script::Script(RuntimeIPtr runtime, BufferPtr bytecode, std::any userContext) :
+	inline Script::Script(RuntimeIPtr runtime, BufferPtr bytecode, Any userContext) :
 		m_runtime(runtime),
 		m_userContext(userContext),
 		m_finished(false),